# Core: Prevent infinite rerendering caused by comparison by reference